+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to thuisbezorgd-scraper
 
+### 1.1.3 (November 12, 2017) 
+- Updated moment.js library to 2.19.2
+- Use console.log() for debug logging instead of cli.debug()
+
 ### 1.1.2 (November 12, 2017) 
 - Added extra check to test if login was successful